+++ CHANGES.md
@@ -1,6 +1,7 @@
 ﻿### 3.35.3 (2026-04-24 18:20:00 UTC)
 
 * Fix browse Metacritic cards to new Metacritic layout
+* Fix episode KeyError with mismatch Aired to DVD Order from TheTVDb API when updating show info
 
 
 ### 3.35.2 (2026-04-09 11:00:00 UTC)